backend/pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG system for Finance/Legal document intelligence"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "redis[hiredis]>=5.0.0",
    "openai>=1.60.0",
    "cohere>=5.0.0",
    "llama-parse>=0.5.0",
    "arq>=0.26.0",
    "tiktoken>=0.8.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.18",
    "email-validator>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
